Unify duplicate Values types, better error reporting about unexpected types while merging maps
