+++ plugin/pluginutil/merge_values.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	"github.com/kubernetes-incubator/kube-aws/plugin/pluginmodel"
+	"reflect"
 )
 
 func MergeValues(v pluginmodel.Values, o map[string]interface{}) pluginmodel.Values {
@@ -18,31 +19,34 @@ func MergeValues(v pluginmodel.Values, o map[string]interface{}) pluginmodel.Val
 func merge(x1, x2 interface{}) interface{} {
 	switch x1 := x1.(type) {
 	case map[string]interface{}:
-		x2, ok := x2.(map[string]interface{})
-		if !ok {
-			panic(fmt.Sprintf("cannot merge map[string]interface{} %+v and %+v", x1, x2))
+		switch x2 := x2.(type) {
+		case map[string]interface{}:
+			for k, v2 := range x2 {
+				if v1, ok := x1[k]; ok {
+					x1[k] = merge(v1, v2)
+				} else {
+					x1[k] = v2
+				}
+			}
+			return x1
+		default:
+			panic(fmt.Sprintf("cannot merge %+v(map[string]interface{}) and %+v(%s)", x1, x2, reflect.TypeOf(x2)))
 		}
-		for k, v2 := range x2 {
-			if v1, ok := x1[k]; ok {
-				x1[k] = merge(v1, v2)
-			} else {
+	case map[string]string:
+		switch x2 := x2.(type) {
+		case map[string]string:
+			for k, v2 := range x2 {
 				x1[k] = v2
 			}
+			r := map[string]interface{}{}
+			for k, v := range x1 {
+				r[k] = string(v)
+			}
+			return r
+
+		default:
+			panic(fmt.Sprintf("cannot merge %+v(map[string]string map[string]string) and %+v(%s)", x1, x2, reflect.TypeOf(x2)))
 		}
-		return x1
-	case map[string]string:
-		x2, ok := x2.(map[string]string)
-		if !ok {
-			panic(fmt.Sprintf("cannot merge map[string]string %+v and %+v", x1, x2))
-		}
-		for k, v2 := range x2 {
-			x1[k] = v2
-		}
-		r := map[string]interface{}{}
-		for k, v := range x1 {
-			r[k] = string(v)
-		}
-		return r
 	case nil:
 		panic(fmt.Sprintf("cannot merge nil and %+v", x2))
 	}
